requirements.txt
scrapy==2.13.0
pandas==2.2.3
pymongo==4.13.0
pytz
python-dotenv